Change the ref() value with uName.value
The ref() function creates an object and returns another object. This,
contains all the variables and methods available to the DOM;
To change the ref(), you can't re-assignt it. Instead, use the
uName.value to assign a new value to the variable in the ref().
Vue will pick it up automatically and updates the DOM.