Exposing objects and data
The best way to expose data to the DOM is by creating a reactive()
to create an object and returning the object.

There are some tools that can come in handy: isReactive os isRefs are
tools to check if something is reactive, returning true or false.

toRefs() is a function that receives an object, any object, and turns
all the properties inside of it to reactive data. So you can return the
data itself if the object it belongs was received by toRefs.