using the watch() function
there is a substitute for the watchers: watch().
the watch can be imported from vue and it works like the watch in the
optional componsition.
In this way, you use the watch function with two arguments:
first, the value you want to watch. It can be an array of value if you
want to watch more values. And the second argument is the function that
should be executed.
If you pass an array of value in the first argument, the function (in
the second argument) will receive the value in the sabe order.

Ex.:

value1 = 1 changing to 4
value2 = 2 changing to 5

watch([value1, value2], function(oldvalues, newvalues) {
    console.log(oldvalues[0]) // 1
    console.log(oldvalues[1] // 2;
    console.log(newvalue[0])) // 4;
    console.log(newvalue[1])  // 5;
})