Working with template refs
It is also possible to work with template refs. You can add the ref
attribute in a DOM element and pass a string name to it.
Then, you can 'catch' it with the ref('') function using the same
string name and get its value.

<input ref="example">
const name = ref('example');
example.value.value because one value is from the template ref and the
other value is from the ref() function.